merkle: Move proof construction essentials to another file
